more shared type extraction
